+++ CHANGELOG.md
@@ -28,6 +28,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
   - This option requires a calibration step that may take a few seconds.
 - Add hardware clock control and diagnostic utility `tstool`.
 
+### Changed
+
+- Default control socket moved to /run/sfptpd/control-v1.sock (SWPTP-1551)
+
 ## [3.8.1.1004] - 2025-02-14
 
 ### Added